Move target checking for `#[lang]` to the attribute parser
