+++ tests/ui/error-codes/E0264.rs
@@ -1,8 +1,8 @@
 #![feature(lang_items)]
 
 extern "C" {
-    #[lang = "copy"]
-    fn copy(); //~ ERROR E0264
+    #[lang = "copy"] //~ ERROR E0264
+    fn copy();
 }
 
 fn main() {}
