+++ tests/ui/panic-handler/panic-handler-wrong-location.stderr
@@ -1,11 +1,12 @@
-error[E0718]: `panic_impl` lang item must be applied to a function
+error: `#[panic_handler]` function required, but not found
+
+error: `#[panic_handler]` attribute cannot be used on statics
   --> $DIR/panic-handler-wrong-location.rs:6:1
    |
 LL | #[panic_handler]
-   | ^^^^^^^^^^^^^^^^ attribute should be applied to a function, not a static
-
-error: `#[panic_handler]` function required, but not found
+   | ^^^^^^^^^^^^^^^^
+   |
+   = help: `#[panic_handler]` can only be applied to functions
 
 error: aborting due to 2 previous errors
 
-For more information about this error, try `rustc --explain E0718`.
